moving username and trim to openai and discord
The message history logic should not need to know how to convert user IDs into usernames.

Additionally, with the knowledge that OpenAI prompt length is based on encoded tokens, not character length, the task of trimming history is too complex for message history logic to handle alone.

WIP: Using the GPT2 encoder (which GPT3 also uses) to calculate prompt length in the openai client. Shifting username and transcript creation to the discord bot module.